% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{browse_np}
\alias{browse_np}
\title{Browse national park web page}
\usage{
browse_np(np)
}
\arguments{
\item{np}{national park name.
Regular expressions and partial matches are valid.}
}
\description{
Browse national park web page
}
\examples{
\dontrun{
browse_np("Shiretoko")
browse_np("Daisetsu")
}
}
