% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_area.R
\name{np_area_meshed}
\alias{np_area_meshed}
\title{Inside of National Park Meshes}
\usage{
np_area_meshed(d, mesh_size = c("10km", "1km"))
}
\arguments{
\item{d}{data (national park polygon)}

\item{mesh_size}{Which mesh size to output. Supports 10km and 1km.}
}
\description{
Inside of National Park Meshes
}
\examples{
\dontrun{
data(np, package = "np")
np[11, ] \%>\% np_area_meshed(mesh_size = "10km")
}
}
