% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{np_information}
\alias{np_information}
\title{Japanese National Park Information}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34 rows and 8 columns.}
\usage{
np_information
}
\description{
Japan National Park Data.
}
\examples{
np_information
}
\keyword{datasets}
