% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{tidy_np_visitor}
\alias{tidy_np_visitor}
\title{read and tidy up national park visitor data}
\usage{
tidy_np_visitor(path, range, col_names)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{range}{A cell range to read from, as described in \link[readxl]{cell-specification}.
Includes typical Excel ranges like "B3:D87", possibly including the sheet
name like "Budget!B2:G14", and more. Interpreted strictly, even if the
range forces the inclusion of leading or trailing empty rows or columns.
Takes precedence over \code{skip}, \code{n_max} and \code{sheet}.}

\item{col_names}{\code{TRUE} to use the first row as column names, \code{FALSE} to get
default names, or a character vector giving a name for each column. If user
provides \code{col_types} as a vector, \code{col_names} can have one entry per
column, i.e. have the same length as \code{col_types}, or one entry per
unskipped column.}
}
\description{
read and tidy up national park visitor data
}
