% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{np}
\alias{np}
\title{Japanese National Park Information Dataset}
\format{A `sf` object contains 34 rows 2 variables:
\itemize{
  \item{name}
  \item{geometry}
}}
\source{
Natural Environmental Information GIS (Biodiversity Center of Japan, Ministry of the Environment.)

\url{http://www.biodic.go.jp/trialSystem/EN/info/nps.html}

This data is modified from GIS data that national park area by Shinya Uryu.
}
\usage{
np
}
\value{
\item{np}{a sf}
}
\description{
Japan National Park Area
}
\details{
National Parks represent outstanding natural landscapes of Japan. The park lands are designated
and managed by the Country (MOE) in accordance with the Natural Parks Law.
As of February 2018, 34 areas have been designated as the National Parks.
The data includes names and classification attributes such as Special Protection Zones,
Class I Special Zones, etc. Please see \url{https://www.env.go.jp/en/nature/nps/park/} the description of each zone.
}
\examples{
\dontrun{
np
}
}
\keyword{datasets}
