% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_area.R
\name{np_area_meshed}
\alias{np_area_meshed}
\title{Inside of National Park Meshes}
\usage{
np_area_meshed(df, mesh_size = c(1, 10))
}
\arguments{
\item{df}{data frame (national park polygon)}

\item{mesh_size}{Which mesh size to output. Supports 10 and 1km.}
}
\description{
Inside of National Park Meshes
}
\examples{
\dontrun{
data(np, package = "np")
np[11, ] \%>\% np_area_meshed(mesh_size = 10)
}
}
